Move Rand backend trait to random::backend module
